LayoutTests/platform/ios-simulator/editing/pasteboard/paste-text-010-expected.txt

There was a bug when pasting at the end of the block. The content was inserted at the start
the block instead of the end. This tests the insert-at-end case.
| "
"
| <div>
|   class="editing"
|   id="test"
|   "There is a tide in the affairs of men."
|   <br>
|   <br>
|   <font>
|     size="3"
|     <span>
|       style="background-color: rgba(255, 255, 255, 0);"
|       "of men."
|       "of men.<#selection-caret>"
| "
"
